+++ cloudflarectl/main.go
@@ -1,14 +1,10 @@
 package main
 
 import (
-	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"sort"
-	"strings"
 
-	cloudflare "github.com/cloudflare/cloudflare-go"
 	"github.com/urfave/cli"
 )
 
@@ -92,52 +88,4 @@ func main() {
 
 	app.Run(os.Args)
 
-}
-
-func clearCache(apiKey string, email string, files []string) error {
-	for i := range files {
-		files[i] = "https://assets.honestbee.com/" + files[i]
-	}
-	purgeCacheRequest := cloudflare.PurgeCacheRequest{
-		Files: files,
-	}
-
-	// Construct a new API object
-	api, err := cloudflare.New(apiKey, email)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Fetch the zone ID
-	id, err := api.ZoneIDByName("honestbee.com") // Assuming example.com exists in your Cloudflare account already
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	purgeCacheRespone, err := api.PurgeCache(id, purgeCacheRequest)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	_, err = fmt.Printf("status: %v", purgeCacheRespone.Response.Success)
-	return err
-}
-
-func parseFile(fileName string) ([]string, error) {
-	file, err := ioutil.ReadFile(fileName)
-	if err != nil {
-		return nil, err
-	}
-	var fileSlice stringsAlias
-	fileSlice = strings.Split(string(file), "\n")
-	return fileSlice.removeEmpty(), nil
-}
-
-func (s stringsAlias) removeEmpty() []string {
-	for i, element := range s {
-		if element == "" {
-			s = append(s[:i], s[i+1:]...)
-		}
-	}
-	return s
 }
